added user and facility crud

+++ srs_intake/facilities/routes.py
@@ -1,16 +1,26 @@
 from flask import render_template, url_for, flash, redirect, request, abort, Blueprint
 from flask_login import current_user, login_required
 from srs_intake import db
-from srs_intake.models import Facility
+from srs_intake.models import Facility, User, Referral
 from srs_intake.facilities.forms import FacilityForm
 
 
 facilities = Blueprint('facilities', __name__)
 
 
+@facilities.route("/facility/list")
+@login_required
+def list_facilities():
+    user = User.query.get(current_user.id)
+    if user.role == 'admin':      
+        facilities = Facility.query.all()
+        return render_template('facilities/facilities_list.html', title="Facilities", facilities=facilities)
+    else:
+        abort(403)
+
 
 @facilities.route("/facility/new", methods=['GET', 'POST'])
-# @login_required
+@login_required
 def new_facility():
     form = FacilityForm()
     if form.validate_on_submit():
@@ -23,18 +33,18 @@ def new_facility():
 
 
 @facilities.route("/facility/<int:facility_id>")
-# @login_required
+@login_required
 def facility(facility_id):
     facility = Facility.query.get_or_404(facility_id)
     return render_template('facilities/facility.html', title=f"{facility.name}", facility=facility)
 
 
 @facilities.route("/facility/<int:facility_id>/update", methods=['GET', 'POST'])
-# @login_required
+@login_required
 def update_facility(facility_id):
     facility = Facility.query.get_or_404(facility_id)
-    # if referral.facility_id != current_user.facility_id or current_user.role != 'admin':
-    #     abort(403)
+    if current_user.role != 'admin':
+         abort(403)
     form = FacilityForm()
     if form.validate_on_submit():
         facility.name=form.name.data
@@ -58,11 +68,11 @@ def update_facility(facility_id):
     return render_template('facilities/crud_facility.html', title='Update Facility', form=form)
 
 @facilities.route("/facility/<int:facility_id>/delete", methods=['POST'])
-# @login_required
+@login_required
 def delete_facility(facility_id):
     facility = Facility.query.get_or_404(facility_id)
-    # if referral.facility_id != current_user.facility_id or current_user.role != 'admin':
-    #     abort(403)
+    if current_user.role != 'admin':
+        abort(403)
     db.session.delete(facility)
     db.session.commit()
     flash('The Facility was deleted successfully.', 'success')
